scraping/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.21.3
azure-storage-blob==12.24.0
pandas==2.2.3
pytz==2024.2
requests==2.32.3
python-dotenv==1.0.1
